Refactor: [BADA-82] enum 패키지 구조 변경 #40
